Move internal/mapping to mapping/internal.
