+++ cartographer/mapping/2d/local_trajectory_builder_options_2d.cc
@@ -16,10 +16,10 @@
 
 #include "cartographer/mapping/2d/local_trajectory_builder_options_2d.h"
 
-#include "cartographer/internal/mapping/motion_filter.h"
 #include "cartographer/mapping/2d/scan_matching/ceres_scan_matcher_2d.h"
 #include "cartographer/mapping/2d/scan_matching/real_time_correlative_scan_matcher_2d.h"
 #include "cartographer/mapping/2d/submap_2d.h"
+#include "cartographer/mapping/internal/motion_filter.h"
 #include "cartographer/sensor/voxel_filter.h"
 
 namespace cartographer {
