+++ cartographer/mapping/2d/scan_matching/ceres_scan_matcher_2d.cc
@@ -22,10 +22,10 @@
 #include "Eigen/Core"
 #include "cartographer/common/ceres_solver_options.h"
 #include "cartographer/common/lua_parameter_dictionary.h"
-#include "cartographer/internal/mapping/2d/scan_matching/occupied_space_cost_function_2d.h"
 #include "cartographer/mapping/2d/probability_grid.h"
 #include "cartographer/mapping/2d/scan_matching/rotation_delta_cost_functor_2d.h"
 #include "cartographer/mapping/2d/scan_matching/translation_delta_cost_functor_2d.h"
+#include "cartographer/mapping/internal/2d/scan_matching/occupied_space_cost_function_2d.h"
 #include "cartographer/transform/transform.h"
 #include "ceres/ceres.h"
 #include "glog/logging.h"
