+++ cartographer/mapping/3d/pose_graph/optimization_problem_3d.cc
@@ -30,12 +30,12 @@
 #include "cartographer/common/make_unique.h"
 #include "cartographer/common/math.h"
 #include "cartographer/common/time.h"
-#include "cartographer/internal/mapping/3d/acceleration_cost_function_3d.h"
-#include "cartographer/internal/mapping/3d/rotation_cost_function_3d.h"
 #include "cartographer/mapping/3d/imu_integration.h"
 #include "cartographer/mapping/3d/pose_graph/landmark_cost_function_3d.h"
 #include "cartographer/mapping/3d/pose_graph/spa_cost_function_3d.h"
 #include "cartographer/mapping/3d/rotation_parameterization.h"
+#include "cartographer/mapping/internal/3d/acceleration_cost_function_3d.h"
+#include "cartographer/mapping/internal/3d/rotation_cost_function_3d.h"
 #include "cartographer/mapping/pose_graph/ceres_pose.h"
 #include "cartographer/transform/timestamped_transform.h"
 #include "cartographer/transform/transform.h"
