+++ cartographer/mapping/internal/motion_filter_test.cc
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "cartographer/internal/mapping/motion_filter.h"
+#include "cartographer/mapping/internal/motion_filter.h"
 
 #include "cartographer/common/lua_parameter_dictionary_test_helpers.h"
 #include "gmock/gmock.h"
